Rename java_card_config.ts config.ts
In prep for addition of VVSG2 data protection config
