+++ libs/auth/src/config.ts
